moved queue and stack to util file
