+++ graph_adventure/adv.py
@@ -1,6 +1,7 @@
 from room import Room
 from player import Player
 from world import World
+from util import Stack, Queue #importing stack and Queue from util file
 
 import random
 
@@ -19,37 +20,14 @@ world.loadGraph(roomGraph)
 world.printRooms()
 player = Player("Name", world.startingRoom)
 
-class Stack():
-    def __init__(self):
-        self.stack = []
-    def push(self,value):
-        self.stack.append(value)
-    def pop(self):
-        if self.size() > 0:
-            return self.stack.pop()
-        else:
-            return None
-    def size(self):
-        return len(self.stack)
-
-class Queue():
-    def __init__(self):
-        self.queue = []
-    def enqueue(self, value):
-        self.queue.append(value)
-    def dequeue(self):
-        if self.size() > 0:
-            return self.queue.pop(0)
-        else:
-            return None
-    def size(self):
-        return len(self.queue)
         
-
+# commands player.currentRoom.id, player.currentRoom.getExits() and player.travel(direction) useful
 
 # FILL THIS IN
-traversalPath = ['n', 's']
+# traversalPath = ['n', 's'] example path
 
+traversalPath = [] # starts empty, need to go to every room
+visited = {}
 
 # TRAVERSAL TEST
 visited_rooms = set()
